Create ResumeForm component
