+++ justTalent/src/app/candidate/candidate.component.ts
@@ -1,16 +1,16 @@
 import { Component, OnInit } from '@angular/core';
-import { User } from '../interfaces/user';
 import { UserService } from '../services/user.service';
 import { Resume } from '../interfaces/resume';
 import { ResumeService } from '../services/resume.services';
 import { CommonModule } from '@angular/common';
-import { FormGroup, ReactiveFormsModule, FormBuilder } from '@angular/forms';
+import { ReactiveFormsModule } from '@angular/forms';
 import { ResumeComponent } from '../resume/resume.component';
+import { ResumeFormComponent } from '../resume-form/resume-form.component';
 
 @Component({
   selector: 'app-candidate',
   standalone: true,
-  imports: [CommonModule, ReactiveFormsModule, ResumeComponent],
+  imports: [CommonModule, ReactiveFormsModule, ResumeComponent, ResumeFormComponent],
   templateUrl: './candidate.component.html',
   styleUrl: './candidate.component.css'
 })
@@ -19,66 +19,33 @@ export class CandidateComponent implements OnInit{
   resume: Resume | undefined;
 
   isEditMode:boolean = false;
-  // Non-null assertion operator used here
-  resumeForm!: FormGroup;
+
 
   constructor(
     private userService: UserService,
     private resumeService: ResumeService,
-    private fb:FormBuilder
   ) {}
 
   ngOnInit(): void {
     this.getCurrentUserDetails();
     this.getCurrentUserResume();
-    this.initializeForm();
 
   }
+
+
 
   parseDate(dateString: string): Date | null {
     return dateString ? new Date(dateString) : null;
   }
 
-  // EDIT BUTTON TOGGLER
   toggleEditMode(): void {
     this.isEditMode = !this.isEditMode;
-    if (this.isEditMode && this.resume) {
-      this.populateForm(this.resume); // Update form with resume data
-    }
   }
 
-  // Populate form with resume data
-  populateForm(resume: Resume): void {
-    this.resumeForm.patchValue({
-      title: resume.title,
-      summary: resume.summary,
-      // Populate other form controls as needed
-    });
-  }
-
-  //FORM SUBMIT HANDLER
-  onSubmit(): void {
-    // Check if the form is valid before proceeding
-    if (this.resumeForm.valid) {
-      // Update the resume object with form values
-      const updatedResume= {
-        ...this.resume,
-        ...this.resumeForm.value
-      };
-      this.resumeService.updateResume(updatedResume).subscribe(()=>{
-        this.resume = updatedResume;
-        this.toggleEditMode();
-      })
-
-    }
-  }
-
-  // Initialize the form with default values
-  initializeForm(): void {
-    this.resumeForm = this.fb.group({
-      title: [this.resume?.title || ''],
-      summary: [this.resume?.summary || ''],
-      // Add other form controls as needed
+  handleFormSubmit(updatedResume: Resume): void {
+    this.resumeService.updateResume(updatedResume).subscribe(() => {
+      this.resume = updatedResume;
+      this.toggleEditMode();
     });
   }
 
@@ -93,7 +60,6 @@ export class CandidateComponent implements OnInit{
     this.resumeService.getCurrentUserResume()
       .subscribe((resume: Resume) => {
         this.resume = resume;
-        console.log(resume, '🚨')
       })
   }
 }
