+++ justTalent/src/app/resume-form/resume-form.component.html
@@ -1,0 +1,10 @@
+<form [formGroup]="resumeForm" (ngSubmit)="onSubmit()">
+  <label for="title">Title</label>
+  <input id="title" formControlName="title" type="text">
+
+  <label for="summary">Summary</label>
+  <textarea id="summary" formControlName="summary"></textarea>
+
+  <button type="submit">Update</button>
+  <button type="button" (click)="cancelEditMode()">Cancel</button>
+</form>
